Se modularizo  incripciones  de competidores , modulo cajer y reportes del administrador
