+++ frontend/src/App.jsx
@@ -3,15 +3,15 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Navbar from "./components/Navbar"; // Importamos la barra de navegación
 import Login from "./components/Login";
 import InicioAdmin from "./components/InicioAdmin";
-import TablaArea from "./components/TablaArea";
+import TablaArea from "./components/areas/TablaArea.jsx";
 import Disciplinas from "./components/Disciplinas.jsx";
 import RegistrarTutores from "./components/RegistrarTutores.jsx";
 import RegistroUsuario from "./components/RegistroUsuario.jsx";
-import RegistroCompetidores from "./components/RegistroCompetidores.jsx";
-import FormularioInscripcion from "./components/FormularioInscripcion.jsx";
+import RegistroCompetidores from "./components/InscripcionCompetidor/RegistroCompetidores.jsx";
+import FormularioInscripcion from "./components/InscripcionCompetidor/FormularioInscripcion.jsx";
 import Pagos from "./components/Pagos.jsx";
 import DetallePago from './components/DetallePago';
-import Reportes from "./components/Reportes.jsx";
+import Reportes from "./components/reportes/Reportes.jsx";
 import HistorialConvocatorias from "./components/HistorialConvocatorias.jsx";
 import Convocatorias from "./components/Convocatorias/Convocatorias.jsx";
 import SolicitudesTutoria from "./components/SolicitudesTutoria.jsx";
