+++ frontend/src/components/areas/ModalConfirmacionEliminar.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../styles/ModalConfirmacionEliminar.css';
+import '../../styles/areas/ModalConfirmacionEliminar.css';
 
 const ModalConfirmacionEliminar = ({ mostrar, cerrar, confirmar, nombreArea }) => {
   if (!mostrar) return null;
