+++ frontend/src/components/cajero/ValidadorPagos.jsx
@@ -1,85 +1,17 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import '../../styles/cajero/ValidadorPagos.css';
-import { Search, RotateCcw, CheckCircle } from 'lucide-react';
-import Swal from 'sweetalert2';
+import { Search, RotateCcw } from 'lucide-react';
+import TablaBoletas from './TablaBoletas';
+import { useValidadorPagos } from '../../hooks/useValidadorPagos';
 import { mostrarConfirmacionPago } from './ConfirmacionPagoModal';
-
 const ValidadorPagos = () => {
-  const [criterio, setCriterio] = useState('codigo');
-  const [termino, setTermino] = useState('');
-  const [boletas, setBoletas] = useState([]);
-  const [boletasOriginal, setBoletasOriginal] = useState([]);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    setTimeout(() => {
-      const datosSimulados = [
-        {
-          codigo: 'BOL-2024-001',
-          competidor: 'Lidia Veizaga',
-          ci: '4567890',
-          monto: 15.0,
-          fecha: '2025-04-01',
-        },
-        {
-          codigo: 'BOL-2024-002',
-          competidor: 'Juan Pérez',
-          ci: '1234567',
-          monto: 20.0,
-          fecha: '2025-04-02',
-        }
-      ];
-      setBoletas(datosSimulados);
-      setBoletasOriginal(datosSimulados);
-    }, 1000);
-  }, []);
-
-  const handleBuscar = () => {
-    const t = termino.trim().toLowerCase();
-    const filtradas = boletasOriginal.filter((b) => {
-      if (criterio === 'codigo') return b.codigo.toLowerCase().includes(t);
-      if (criterio === 'nombre') return b.competidor.toLowerCase().includes(t);
-      if (criterio === 'ci') return b.ci.includes(t);
-      return false;
-    });
-    setBoletas(filtradas);
-  };
-
-  const handleResetear = () => {
-    setTermino('');
-    setBoletas(boletasOriginal);
-  };
-
-  const confirmarValidacion = async (boleta) => {
-    const confirmado = await mostrarConfirmacionPago(boleta);
-    if (!confirmado) return;
-
-    try {
-      await fetch('/api/validar-pago', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ codigo: boleta.codigo })
-      });
-
-      const actualizadas = boletas.filter(b => b.codigo !== boleta.codigo);
-      setBoletas(actualizadas);
-      setBoletasOriginal(actualizadas);
-
-      Swal.fire({
-        icon: 'success',
-        title: '¡Pago Validado!',
-        text: 'El pago fue registrado correctamente.',
-        confirmButtonColor: '#4caf50'
-      });
-    } catch (error) {
-      Swal.fire({
-        icon: 'error',
-        title: 'Error',
-        text: 'No se pudo validar el pago. Intente nuevamente.',
-        confirmButtonColor: '#e53935'
-      });
-    }
-  };
+  const {
+    criterio, setCriterio,
+    termino, setTermino,
+    boletas, error,
+    handleBuscar, handleResetear, confirmarValidacion,
+    esTerminoValido
+  } = useValidadorPagos();
 
   return (
     <div className="validador-container">
@@ -103,7 +35,12 @@ const ValidadorPagos = () => {
               type="text"
               placeholder="Ej: BOL-2025-001"
               value={termino}
-              onChange={(e) => setTermino(e.target.value)}
+              onChange={(e) => {
+                const valor = e.target.value;
+                if (esTerminoValido(valor)) {
+                  setTermino(valor);
+                }
+              }}
             />
           </div>
 
@@ -119,42 +56,11 @@ const ValidadorPagos = () => {
 
       <div className="validador-tabla">
         <h3>Boletas Pendientes</h3>
-        {!error && boletas.length > 0 && (
-          <table>
-            <thead>
-              <tr>
-                <th>CÓDIGO</th>
-                <th>COMPETIDOR</th>
-                <th>CI</th>
-                <th>MONTO</th>
-                <th>FECHA</th>
-                <th>ACCIÓN</th>
-              </tr>
-            </thead>
-            <tbody>
-              {boletas.map((b, i) => (
-                <tr key={i}>
-                  <td>{b.codigo}</td>
-                  <td>{b.competidor}</td>
-                  <td>{b.ci}</td>
-                  <td>Bs. {b.monto.toFixed(2)}</td>
-                  <td>{b.fecha}</td>
-                  <td>
-                    <button className="btn-validar" onClick={() => confirmarValidacion(b)}>
-                      Validar Pago
-                    </button>
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
+        {!error ? (
+          <TablaBoletas boletas={boletas} onConfirmar={confirmarValidacion} />
+        ) : (
+          <p style={{ color: 'red' }}>{error}</p>
         )}
-        {!error && boletas.length === 0 && (
-          <p style={{ textAlign: 'center', marginTop: '1rem' }}>
-            No se encontraron resultados.
-          </p>
-        )}
-        {error && <p style={{ color: 'red' }}>{error}</p>}
       </div>
     </div>
   );
