+++ frontend/src/forms/validadorPagosValidator.js
@@ -1,0 +1,5 @@
+// src/validators/validadorPagosValidator.js
+
+export const esTerminoValido = (texto) => {
+    return texto.length <= 20;
+  };
