+++ frontend/src/styles/HistorialConvocatorias.css
@@ -1,34 +1,43 @@
-/* Reset básico para evitar márgenes inesperados */
-body {
+/* Reset básico */
+html, body, #root {
+  height: 100%;
   margin: 0;
   padding: 0;
   font-family: 'Comfortaa', sans-serif;
   background: #f1f6f9;
 }
 
-.historial-container {
-  flex: 1;
-  padding: 2rem;
-  padding-bottom: 0;
-  max-width: 1300px !important;     /* ajusta el ancho máximo de las tarjetas */
-  margin: 0 auto;        /* centra el contenedor */
+/* Contenedor principal */
+.historial-wrapper {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
 }
 
-.titulo {
+.historial-contenedor {
+  flex: 1;
+  padding: 2rem;
+  max-width: 3000px;
+  margin: 0 auto;
+}
+
+/* Título y subtítulo */
+.historial-titulo {
   text-align: center;
   font-size: 2rem;
   margin-bottom: 0.3rem;
   color: #111;
 }
 
-.subtitulo {
+.historial-subtitulo {
   text-align: center;
   font-size: 1rem;
   color: #555;
   margin-bottom: 2rem;
 }
 
-.card-convocatoria {
+/* Card de convocatoria */
+.historial-card {
   display: flex;
   align-items: center;
   background: #fff;
@@ -36,12 +45,12 @@ body {
   margin-bottom: 3rem;
   box-shadow: 0 3px 7px rgba(0, 0, 0, 0.08);
   overflow: hidden;
-  min-height: 140px; /* Aumenta el alto de las tarjetas */
+  min-height: 140px;
 }
 
-.barra-anio {
+.historial-barra-anio {
   width: 80px;
-  min-height: 140px; /* Coincide con el nuevo alto de tarjeta */
+  min-height: 140px;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -50,19 +59,19 @@ body {
   font-size: 1.3rem;
 }
 
-.contenido {
+.historial-card-contenido {
   flex-grow: 1;
   padding: 1rem 1.5rem;
 }
 
-.nombre {
+.historial-nombre {
   margin: 0;
   font-size: 1.1rem;
   font-weight: 500;
   color: #222;
 }
 
-.fechas {
+.historial-fechas {
   display: flex;
   gap: 2rem;
   margin-top: 0.5rem;
@@ -70,17 +79,18 @@ body {
   color: #444;
 }
 
-.fecha {
+.historial-fecha {
   display: flex;
   align-items: center;
   gap: 0.5rem;
 }
 
-.icono {
-  color: #4a90e2;
+.historial-icono {
+  color: red; /* color rojo para los íconos */
 }
 
-.estado {
+/* Estados */
+.historial-estado {
   margin-right: 1rem;
   font-size: 0.75rem;
   padding: 0.4rem 0.7rem;
@@ -90,49 +100,52 @@ body {
   white-space: nowrap;
 }
 
-.estado.abierto {
+.historial-estado.abierto {
   background-color: #e3f9ed;
   color: #28a745;
 }
 
-.estado.finalizado {
+.historial-estado.finalizado {
   background-color: #e9e9ec;
   color: #5a5a5a;
 }
 
-.footer {
-  width: 100vw;
+/* Footer */
+.historial-footer {
+  width: 100%;
   background-color: #0e1421;
   color: white;
   display: flex;
   justify-content: space-between;
-  padding: 1rem 2rem; /* Altura más reducida */
+  padding: 1rem 2rem;
   font-size: 0.9rem;
   flex-wrap: wrap;
-  margin: 0;
 }
 
-.footer-info {
+.historial-footer-info {
   max-width: 70%;
   line-height: 1.6;
 }
 
-.footer-logo {
+.historial-footer-logo {
   text-align: right;
 }
 
-.footer-logo p {
+.historial-footer-logo p {
   font-weight: bold;
   margin-bottom: 0.5rem;
 }
 
-.footer-icons {
+.historial-footer-icons {
   width: 120px;
   height: auto;
   margin-top: 0.5rem;
 }
-html, body, #root {
-  height: 100%;
-  margin: 0;
-  padding: 0;
+
+/* Mensaje de carga (si lo dejas activado) */
+.historial-cargando {
+  text-align: center;
+  font-size: 1.2rem;
+  color: #555;
+  margin-top: 2rem;
 }
