+++ frontend/src/styles/InscripcionCompetidor/RegistroCompetidores.css
