+++ frontend/src/styles/areas/ModalConfirmacionEliminarCategoria.css
