+++ frontend/src/styles/areas/ModalNuevaArea.css
