+++ frontend/src/styles/areas/ModalNuevaCategoria.css
