+++ frontend/src/styles/areas/TablaArea.css
